# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: FloatTranslate Local API
  version: 1.1.0
  description: >-
    Local loopback API contract between the Vue frontend and Go backend.
    Business endpoints require a per-run bearer token generated by the Tauri host.
servers:
  - url: http://127.0.0.1:{port}/api/v1
    variables:
      port:
        default: "0"
security:
  - LocalSessionBearer: []
paths:
...
  /translations:
    post:
      operationId: createTranslation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TranslationRequest" }
      responses:
        "200":
          description: Translation result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TranslationResponse" }
        default:
          $ref: "#/components/responses/Error"
  /translations/{id}:
    get:
      operationId: getTranslation
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Saved translation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TranslationResponse" }
  /translations/{id}/retranslate:
    post:
      operationId: retranslate
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Fresh result bypassing cache
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TranslationResponse" }
...
components:
  securitySchemes:
    LocalSessionBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-local-session-token
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Standard error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code: { type: string }
            message: { type: string }
            retryable: { type: boolean }
            details: { type: object, additionalProperties: true }
...
    TranslationRequest:
      type: object
      required: [text]
      properties:
        text: { type: string, minLength: 1 }
        force_kind: { type: string, enum: [word, text] }
        bypass_cache: { type: boolean, default: false }
    TranslationResponse:
      type: object
      required: [translation_id, kind, source, result, model, created_at]
      properties:
        translation_id: { type: string }
        kind: { type: string, enum: [word, text] }
        source: { type: string, enum: [model, cache] }
        result:
          oneOf:
            - $ref: "#/components/schemas/WordTranslation"
            - $ref: "#/components/schemas/TextTranslation"
        model: { type: string }
        created_at: { type: string, format: date-time }
    WordTranslation:
      type: object
      required: [word, lemma, phonetic_uk, phonetic_us, parts_of_speech, synonyms, inflections]
      properties:
        word: { type: string }
        lemma: { type: string }
        phonetic_uk: { type: string }
        phonetic_us: { type: string }
        parts_of_speech:
          type: array
          items:
            type: object
            required: [part, meanings]
            properties:
              part: { type: string }
              meanings: { type: array, items: { type: string } }
        synonyms: { type: array, items: { type: string } }
        inflections: { type: array, items: { type: string } }
    TextTranslation:
      type: object
      required: [source_markdown, translated_markdown, segments]
      properties:
        source_markdown: { type: string }
        translated_markdown: { type: string }
        segments:
          type: array
          items:
            type: object
            required: [source, translation]
            properties:
              source: { type: string }
              translation: { type: string }
...